draft of compareTo and equals rule by JSL

+++ esql-data/src/main/java/esql/data/ConvertValueToObject.java
@@ -19,24 +19,24 @@ public class ConvertValueToObject implements Function<Value, Object> {
             case TYPE_TIMESTAMP:
                 return ((ValueDateTime)value).toTemporal();
             case TYPE_BYTE:
-                return ((ValueNumber)value).byteValue(); //auto-boxer
+                return Byte.valueOf( ((ValueNumber)value).byteValue() ); //auto-boxer
             case TYPE_UBYTE:
             case TYPE_SHORT:
-                return ((ValueNumber)value).shortValue(); //auto-boxer
+                return Short.valueOf( ((ValueNumber)value).shortValue() ); //auto-boxer
             case TYPE_USHORT:
             case TYPE_INT:
-                return ((ValueNumber)value).intValue(); //auto-boxer
+                return Integer.valueOf( ((ValueNumber)value).intValue() ); //auto-boxer
             case TYPE_UINT:
             case TYPE_LONG:
-                return ((ValueNumber)value).longValue(); //auto-boxer
+                return Long.valueOf( ((ValueNumber)value).longValue() ); //auto-boxer
             case TYPE_ULONG:
                 return ((ValueNumber)value).bigIntValue(); //real object
             case TYPE_DECIMAL:
                 return ((ValueNumber)value).decimalValue(); //real object
             case TYPE_FLOAT:
-                return ((ValueNumber)value).floatValue(); //auto-boxer
+                return Float.valueOf( ((ValueNumber)value).floatValue() ); //auto-boxer
             case TYPE_DOUBLE:
-                return ((ValueNumber)value).doubleValue(); //auto-boxer
+                return Double.valueOf( ((ValueNumber)value).doubleValue() ); //auto-boxer
             case TYPE_BOOLEAN:
                 return ((ValueBoolean)value).booleanValue() ? Boolean.TRUE : Boolean.FALSE; //auto-boxer
             case TYPE_STRING:
